updated: ui - designsystem module decomposition
