+++ core/design-system/src/main/java/dev/enesky/core/design_system/components/SwipeRefresh.kt
@@ -13,8 +13,8 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.unit.dp
+import dev.enesky.core.design_system.annotation.PreviewUiMode
 import dev.enesky.core.design_system.theme.DoodleTheme
-import dev.enesky.core.ui.annotation.PreviewUiMode
 
 @OptIn(ExperimentalMaterialApi::class)
 @Composable
