+++ core/ui/src/main/java/dev/enesky/core/ui/components/home/TopAnimePreview.kt
@@ -19,8 +19,8 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import dev.enesky.core.design_system.R
-import dev.enesky.core.design_system.common.DoodleImagePlaceholder
-import dev.enesky.core.design_system.common.DoodleNetworkImage
+import dev.enesky.core.design_system.components.DoodleImagePlaceholder
+import dev.enesky.core.design_system.components.DoodleNetworkImage
 import dev.enesky.core.design_system.theme.DoodleTheme
 import dev.enesky.core.domain.models.Anime
 import dev.enesky.core.domain.models.placeholderAnime
