+++ feature/login/src/main/java/dev/enesky/feature/login/signup/SignUpScreen.kt
@@ -18,7 +18,7 @@ import dev.enesky.core.common.utils.ObserveAsEvents
 import dev.enesky.core.ui.components.login.EmailAuthComponent
 import dev.enesky.core.ui.components.login.LoginHeader
 import dev.enesky.core.design_system.theme.DoodleTheme
-import dev.enesky.core.ui.annotation.PreviewUiMode
+import dev.enesky.core.design_system.annotation.PreviewUiMode
 import dev.enesky.feature.login.R
 import dev.enesky.feature.login.signup.helpers.SignUpEvents
 import dev.enesky.feature.login.signup.helpers.SignUpUiState
